+++ src/types/auth.ts
@@ -1,6 +1,4 @@
-import { Request } from 'express';
 import { Role } from '../config/prisma';
-import { ParsedQs } from 'qs';
 
 export interface SignUpRequest {
   role: Role;
@@ -17,16 +15,4 @@ export interface LoginRequest {
 export interface TokenResponse {
   accessToken: string;
   refreshToken: string;
-}
-
-export interface AuthenticatedRequest<Params = Record<string, never>, Query = ParsedQs, Body = Record<string, never>>
-  extends Request<Params, any, Body, Query> {
-  payload?: PayloadData;
-}
-
-export interface PayloadData {
-  id: string;
-  sub: string;
-  type: 'access' | 'refresh';
-  iat: number;
 }
