+++ src/types/index.ts
@@ -1,5 +1,20 @@
+import { Request } from 'express';
+import { ParsedQs } from 'qs';
+
 export interface BasicResponse {
   message: string;
+}
+
+export interface PayloadData {
+  id: string;
+  sub: string;
+  type: 'access' | 'refresh';
+  iat: number;
+}
+
+export interface AuthenticatedRequest<Params = Record<string, never>, Query = ParsedQs, Body = Record<string, never>>
+  extends Request<Params, any, Body, Query> {
+  payload?: PayloadData;
 }
 
 export const REDIS_KEY = {
